Make LinearFormParser generate a LinearForm over doubles instead of ints.
Doubles are more general than ints, therefore it is a better temporary
solution until a proper handling of numeric types is implemented.

+++ core/src/main/scala/it/unich/jandom/parsers/LPInvParser.scala
@@ -64,7 +64,7 @@ class LPInvParser(val env: Environment) extends JavaTokenParsers with LinearForm
         }
       }
 
-  private val assignment: Parser[LinearAssignment[Int]] =
+  private val assignment: Parser[LinearAssignment[Double]] =
     (ident <~ ":=") ~ linform ^^ {
       case v ~ lf => LinearAssignment(env.getBindingOrAdd(v), lf)
     }
